Add the open source releases of Containers On The Go
The last version (1.14?) of "Containers On The Go" was regrettably
made non-free. Reportedly, according to some of the reviews from the
time, the code was also obfuscated (I personally did not look at the
code of the non-free version and never upgraded to it). Even more
regrettably, the add-on was recently (mid-2019) removed from the AMO
altogether, though I'm not sure about the reason.

These are the XPIs that I had archived at the time.

The license from 1.1 to 1.10 was MPL-2.0 (based on memory and
partially supported by the archive:

https://web.archive.org/web/20170901095125/https://addons.mozilla.org/en-GB/firefox/addon/containers-on-the-go/

).

The license for 1.11 and 1.13 was GPL-3.0 (fully corroborated by the
archive:

https://web.archive.org/web/20170917103402/https://addons.mozilla.org/en-GB/firefox/addon/containers-on-the-go/versions/

).